requirements.txt
flask
flask-sqlalchemy
flask-login
werkzeug
opencv-python
numpy
pillow
matplotlib
bcrypt
pytz